Change: Move registerServiceWorker into util/
