+++ src/util/registerServiceWorker.js
